requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
pymysql
PyJWT
datetime
flask-login
